+++ HISTORY.rst
@@ -30,7 +30,8 @@ History
   references to ``customer.sources`` to ``customer.legacy_cards`` instead.
   The ``legacy_cards`` attribute already exists in 1.2.0.
 - ``Customer.sources_v3`` is now named ``Customer.sources``.
-
+- A new property ``Customer.payment_methods`` is now available, which allows
+  you to iterate over all of a customer's payment methods (sources then cards).
 
 1.2.1 (2018-07-18)
 ------------------